Move keypoints into core package and update imports
Move keypoints.py into napari_deeplabcut/core (renamed to core.keypoints) and update all imports to reference napari_deeplabcut.core.keypoints. Updated usages across package modules (widgets, ui/color_scheme_display, core/layers, core/trails, ui/labels_and_dropdown) and tests (including a test file rename) to import build_color_cycle(s) and Keypoint types from the new location. Keeps behavior unchanged while consolidating core functionality under the core package.
